+++ web_src/js/modules/fomantic/base.js
