+++ web_src/js/modules/fomantic/checkbox.js
